+++ users/tests/test_views.py
@@ -132,3 +132,30 @@ class AccountTests(APITestCase):
         client.force_authenticate(user=self.test_user)
         response = client.post(reversed_url)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_users_follow_view(self):
+        # test user should be following 0 people at the start
+        user_following_before = self.test_user.following.count()
+        # another_user should have 0 followers at the start
+        user_followers_before = self.another_user.userprofile.followers.count()
+
+        client = APIClient()
+        client.force_authenticate(user=self.test_user)
+        response = client.post('/api/users/praveen/follow/',args=[self.another_user.username])
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        url = 'users-api:following'
+        reversed_url = reverse(url)
+        client = APIClient()
+        client.force_authenticate(user=self.test_user)
+        response = client.get(reversed_url)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        user_following_after = self.test_user.following.count()
+        user_followers_after = self.another_user.userprofile.followers.count()
+
+        self.assertEqual(user_followers_after,user_followers_before + 1)
+        self.assertEqual(user_following_after,user_following_before + 1)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+
